Indicating compressed PANID field in Mac Header as indicated by 15.4-2011 standard page 41. We are assuming that src and dst panid are the same and hence only destination panid is sent and panid compression must be set to 1.
"If both destination and source addressing information is present, the MAC sublayer shall compare the
destination and source PAN identifiers. If the PAN identifiers are identical, the PAN ID Compression field
shall be set to one, and the source PAN identifier shall be omitted from the transmitted frame. If the PAN
identifiers are different, the PAN ID Compression field shall be set to zero, and both Destination PAN
Identifier and Source PAN Identifier fields shall be included in the transmitted frame. If only either the
destination or the source addressing information is present, the PAN ID Compression field shall be set to
zero, and the PAN identifier field of the single address shall be included in the transmitted frame."