FW-767. Add placeholder for a protected resource
The implementation of the RS role of the ACE framework.
RS exposes "/resource1" that should only be accessed over an OSCORE
channel. The security context to establish this channel is communicated
to the RS by the Client by a POST to /token resource, with payload
containing the access token. The access token, encrypted and
authenticated with a key shared between the RS and the AS, contains
OSCORE parameters.